@@ -19,8 +19,7 @@
   "private": true,
   "workspaces": {
     "packages": [
-      "packages/*",
-      "doc-utils"
+      "packages/*"
     ]
   },
   "scripts": {
@@ -28,16 +27,17 @@
     "build": "lerna run build --stream",
     "clean": "lerna run clean --stream",
     "commitmsg": "commitlint -E GIT_PARAMS",
-    "release": "npm run prepare && node scripts/dist/release/index.js",
-    "lerna": "lerna",
     "prepare": "yarn build",
+    "release": "yarn prepare && node scripts/dist/release/index.js",
+    "lerna": "lerna",
+    "lint": "tslint -p tsconfig.lint.json",
     "storybook": "start-storybook -p 9001 -c .storybook",
     "storybook:build": "build-storybook"
   },
   "dependencies": {
     "highlight.js": "^9.13.1",
-    "react": "^16.5.2",
-    "react-dom": "^16.5.2"
+    "react": "^16.6.0",
+    "react-dom": "^16.6.0"
   },
   "devDependencies": {
     "@babel/core": "^7.1.2",
@@ -51,14 +51,14 @@
     "@storybook/addons": "4.0.0",
     "@storybook/react": "4.0.0",
     "@types/chalk": "^2.2.0",
-    "@types/debug": "^0.0.30",
+    "@types/debug": "^0.0.31",
     "@types/execa": "^0.9.0",
     "@types/get-stream": "^3.0.1",
     "@types/glob": "^7.1.0",
     "@types/listr": "^0.13.0",
-    "@types/node": "^10.10.3",
+    "@types/node": "^10.12.0",
     "@types/ora": "^1.3.4",
-    "@types/react": "^16.4.14",
+    "@types/react": "^16.4.18",
     "@types/react-dom": "^16.0.7",
     "@types/storybook-addon-jsx": "^5.4.0",
     "@types/storybook__addon-actions": "^3.4.1",
@@ -69,7 +69,7 @@
     "awesome-typescript-loader": "^5.2.1",
     "babel-core": "7.0.0-bridge.0",
     "babel-loader": "^8.0.4",
-    "css-loader": "^1.0.0",
+    "css-loader": "^1.0.1",
     "debug": "^4.0.1",
     "dependency-graph": "^0.7.2",
     "execa": "^1.0.0",
@@ -80,7 +80,7 @@
     "lerna": "^3.4.0",
     "listr": "^0.14.2",
     "moment": "^2.22.2",
-    "node-sass": "^4.9.3",
+    "node-sass": "^4.9.4",
     "ora": "^3.0.0",
     "prettier": "^1.14.3",
     "react-docgen-typescript-loader": "^3.0.0",
@@ -90,7 +90,7 @@
     "rollup": "^0.66.2",
     "rollup-plugin-auto-external": "^2.0.0",
     "rollup-plugin-commonjs": "^9.1.8",
-    "rollup-plugin-filesize": "^4.0.1",
+    "rollup-plugin-filesize": "^5.0.1",
     "rollup-plugin-node-resolve": "^3.4.0",
     "rollup-plugin-replace": "^2.0.0",
     "rollup-plugin-sourcemaps": "^0.4.2",
@@ -103,7 +103,7 @@
     "tslint": "^5.10.0",
     "tslint-config-kata": "^1.0.5",
     "tslint-config-prettier": "^1.15.0",
-    "tslint-plugin-prettier": "^2.0.0",
+    "tslint-plugin-prettier": "^2.0.1",
     "typescript": "^3.0.3",
     "url-loader": "^1.1.1"
   },
